* changed a lot of code that was still referencing audio
* changed the menu item to be '(de)activate video' with the view of 'video-config.html' so that it does not conflict with the audio menu item

* TODO: there is still some code that is referencing audio specific things that needs to be cleared out. 

The syndication is commented out at the moment.

The only 'plugin' that is setup right now is for a quicktime .mov file.  There is no automatic retrieval of metadata from the video file.  For right now we need to use annotations to have metadata on the video file (this is not set up either, it's static in mov/_videodata.py)