Fix handling of portal status messages for #38.
There were a couple of problems here and the combination was a bit tricky.  The easy part was changing from "?portal_status_message" style messages to IStatusMessage.  The tricky part was the bacause the form action handlers were returning none, that meant the template was being rendered before the redirect.  Aside from just being wasteful, this meant the status messages were being consumed by the template rendering.  Then the redirect renders the actual destination template but by that time there are no portal status messages left to render. This is addressed by returning a non-None value from the action handler in which case the template is not rendered.